ugc/requirements.txt
orjson==3.10.0

aio-pika==9.4.1
kafka-python==2.0.2

flask[async]==3.0.2
flask-swagger-ui==4.11.1
pydantic==2.6.4
pydantic-settings==2.2.1

pyjwt[crypto]==2.8.0
werkzeug==3.0.1

aiokafka==0.10.0
gunicorn==21.2.0
uvicorn==0.27.1
gevent==24.2.1
fast-depends==2.4.2

beanie==1.25.0
pymongo==4.6.3


sentry-sdk[flask]
python-logstash==0.4.8
python-dotenv==0.21.0
